Delete all the database files in a database on drop.

+++ resources/databases.ts
@@ -425,17 +425,23 @@ export function database({ database: database_name, table: table_name }) {
  */
 export async function dropDatabase(database_name) {
 	if (!databases[database_name]) throw new Error('Schema does not exist');
+	let database = databases[database_name];
+	for (const table_name in database) {
+		let table = database[table_name];
+		let root_store = table.primaryStore.rootStore;
+		database_envs.delete(root_store.path);
+		if (root_store.status === 'open') {
+			await root_store.close();
+			await fs.remove(root_store.path);
+		}
+	}
 	if (database_name === 'data') {
 		for (const table_name in tables) {
 			delete tables[table_name];
 		}
 		delete tables[DEFINED_TABLES];
 	}
-	const root_store = database({ database: database_name });
 	delete databases[database_name];
-	database_envs.delete(root_store.path);
-	await root_store.close();
-	await fs.remove(root_store.path);
 }
 
 /**
